Lots of changes, new issues
Cell needs to be able to estimate discrepancy to determine if subdivision is needed. Unfortunately, the way this method is described is reliant on the Marching Cubes algorithm. I will try to experiment with ways to calculate this without falling to a single algorithm, but I may need to give up part of the design and focus on Marching Cubes.