Relicensed to Apache 2.0
From what I can gather, Apache 2.0 includes a weak patent retaliation claim preventing companies from suing other companies or individuals based on code contributed to this library. This protects everyone involved in the use of this library from patent trolls.

If anyone reads this, this is not legal advice, and I am not a lawyer.